safari-600.7-branch/LayoutTests/http/tests/uri/utf8-path.html

<html>
<head>
<meta http-equiv="Content-Type" content="text/html; charset=windows-1251">
</head>
<body>

<script>
    if (window.testRunner)
        testRunner.dumpAsText();
</script>

<p>Test for <a href='https://bugs.webkit.org/show_bug.cgi?id=7461'>bug 7461</a>: 
Always encode the path part of an URI as UTF-8</p>

Test the URL as it is passed to the server. WinIE 6 and Firefox 3.0a results are currently diffrerent:
IE uses the target encoding for the query path, while Firefox uses the page encoding. Also, WinIE doesn't percent-encode
the query path.<br>
"<script id=scr1 src="intercept/print/פאיכ.js?קוי=למי"></script>" (no target charset specified)<br>
"<script id=scr2 charset="windows-1251" src="intercept/print/פאיכ.js?קוי=למי"></script>" (target charset=windows-1251, same as page)<br>
"<script id=scr3 charset="iso-8859-5" src="intercept/print/פאיכ.js?קוי=למי"></script>" (target charset=iso-8859-5)<br><br>

Show the source attribute of the scripts. WinIE 6 doesn't use percent encoding here, while Firefox does.<br>
<script>
 try {
  document.write('"' + document.scripts[1].src + '"<br>');
  document.write('"' + document.scripts[2].src + '"<br>');
  document.write('"' + document.scripts[3].src + '"<br>');
 } catch (ex) {
  document.write('"' + document.getElementById("scr1").src + '"<br>');
  document.write('"' + document.getElementById("scr2").src + '"<br>');
  document.write('"' + document.getElementById("scr3").src + '"<br>');
 }
</script>

</body>
</html>
