third_party/blink/web_tests/platform/generic/http/tests/uri/utf8-path-expected.txt

Test for bug 7461: Always encode the path part of an URI as UTF-8

Test the URL as it is passed to the server. WinIE 6 and Firefox 3.0a results are currently diffrerent: IE uses the target encoding for the query path, while Firefox uses the page encoding. Also, WinIE doesn't percent-encode the query path.
"/uri/intercept/print/%D1%84%D0%B0%D0%B9%D0%BB.js?%F7%E5%E9=%EC%EE%E9" (no target charset specified)
"/uri/intercept/print/%D1%84%D0%B0%D0%B9%D0%BB.js?%F7%E5%E9=%EC%EE%E9" (target charset=windows-1251, same as page)
"/uri/intercept/print/%D1%84%D0%B0%D0%B9%D0%BB.js?%F7%E5%E9=%EC%EE%E9" (target charset=iso-8859-5)

Show the source attribute of the scripts. WinIE 6 doesn't use percent encoding here, while Firefox does.
"http://127.0.0.1:8000/uri/intercept/print/%D1%84%D0%B0%D0%B9%D0%BB.js?%F7%E5%E9=%EC%EE%E9"
"http://127.0.0.1:8000/uri/intercept/print/%D1%84%D0%B0%D0%B9%D0%BB.js?%F7%E5%E9=%EC%EE%E9"
"http://127.0.0.1:8000/uri/intercept/print/%D1%84%D0%B0%D0%B9%D0%BB.js?%F7%E5%E9=%EC%EE%E9"
